refactor(model): move volume and agent normalization to their owning packages
pkg/model holds 88 types and 101 free functions and is imported by 214 of 496
production files, so it couples every layer to one package. AGENTS.md asks for
business rules to live in the package that owns the concept; volume
normalization living next to the volume structs is the clearest violation.

Move the volume normalization family to pkg/volumes, dropping the redundant
package name from each identifier (NormalizeVolumeName becomes
volumes.NormalizeName). The types stay in pkg/model for now; this change moves
behavior, not data.

NormalizeAgentDefinition moves to pkg/projects because it has to. pkg/model
called it internally through NormalizeVolumeMountSpecs, so leaving it behind
would have required pkg/model to import pkg/volumes and close a cycle. Its two
unexported helpers, normalizeCapsetIDs and isJSONObject, have no other caller
and move with it, so nothing in pkg/model had to be exported to make this work.

New edges are storage/configstore, storage/sandboxstore, schedulers, and
agentcompose/adapters to volumes, and projects to volumes. All are in the
existing direction: configstore already imported projects, runs, and
schedulers, and sandboxstore already imported sandboxes. go list confirms
pkg/volumes still depends only on config and model, so no cycle is created.

pkg/model drops from 2580 to 2354 production lines.

Verified with go vet ./pkg/... ./cmd/..., a full golangci-lint run, and
go test over model, volumes, projects, and schedulers. Two pre-existing
macOS-only failures in pkg/volumes (/private/var symlink resolution) reproduce
unchanged on the parent commit.
