+++ pkg/model/agent_model.go
@@ -1,7 +1,6 @@
 package model
 
 import (
-	"encoding/json"
 	"fmt"
 	"strings"
 	"time"
@@ -106,62 +105,6 @@ func NormalizeAgentKind(agent string) string {
 	}
 }
 
-func NormalizeAgentDefinition(item AgentDefinition, assignDefaults bool) (AgentDefinition, error) {
-	item.ID = strings.TrimSpace(item.ID)
-	item.Name = strings.TrimSpace(item.Name)
-	item.Description = strings.TrimSpace(item.Description)
-	item.Provider = NormalizeAgentKind(item.Provider)
-	if item.Provider == "" && assignDefaults {
-		item.Provider = DefaultAgentProvider
-	}
-	item.Model = strings.TrimSpace(item.Model)
-	item.SystemPrompt = strings.TrimSpace(item.SystemPrompt)
-	item.Driver = strings.TrimSpace(item.Driver)
-	item.GuestImage = strings.TrimSpace(item.GuestImage)
-	item.WorkspaceID = strings.TrimSpace(item.WorkspaceID)
-	item.CapsetIDs = normalizeCapsetIDs(item.CapsetIDs)
-	item.ProjectID = strings.TrimSpace(item.ProjectID)
-	item.AgentName = strings.TrimSpace(item.AgentName)
-	item.ConfigJSON = strings.TrimSpace(item.ConfigJSON)
-	if item.ConfigJSON == "" {
-		item.ConfigJSON = "{}"
-	}
-	item.Skills = NormalizeAgentSkills(item.Skills)
-	if item.ID == "" {
-		return AgentDefinition{}, fmt.Errorf("agent definition id is required")
-	}
-	if item.Name == "" {
-		return AgentDefinition{}, fmt.Errorf("agent definition name is required")
-	}
-	if item.Provider == "" {
-		return AgentDefinition{}, fmt.Errorf("agent definition provider is required")
-	}
-	if item.Provider != "codex" && item.Provider != "claude" && item.Provider != "gemini" && item.Provider != "opencode" && item.Provider != "pi" {
-		return AgentDefinition{}, fmt.Errorf("agent definition provider %q is not supported", item.Provider)
-	}
-	if !isJSONObject(item.ConfigJSON) {
-		return AgentDefinition{}, fmt.Errorf("agent definition config_json must be a JSON object")
-	}
-	if item.ProjectID == "" {
-		item.ProjectRevision = 0
-		item.AgentName = ""
-	} else {
-		if item.AgentName == "" {
-			return AgentDefinition{}, fmt.Errorf("managed agent name is required")
-		}
-		if item.ProjectRevision < 0 {
-			return AgentDefinition{}, fmt.Errorf("managed project revision cannot be negative")
-		}
-	}
-	item.EnvItems = NormalizeEnvItems(item.EnvItems)
-	volumes, err := NormalizeVolumeMountSpecs(item.Volumes)
-	if err != nil {
-		return AgentDefinition{}, fmt.Errorf("agent definition volumes: %w", err)
-	}
-	item.Volumes = volumes
-	return item, nil
-}
-
 func NormalizeAgentSkills(skills []AgentSkill) []AgentSkill {
 	if len(skills) == 0 {
 		return nil
@@ -206,31 +149,6 @@ func AgentSkillSource(skill AgentSkill) sources.Source {
 		Password: skill.Password,
 		Token:    skill.Token,
 	}.Normalized()
-}
-
-func normalizeCapsetIDs(ids []string) []string {
-	seen := make(map[string]struct{}, len(ids))
-	out := make([]string, 0, len(ids))
-	for _, id := range ids {
-		id = strings.TrimSpace(id)
-		if id == "" {
-			continue
-		}
-		if _, ok := seen[id]; ok {
-			continue
-		}
-		seen[id] = struct{}{}
-		out = append(out, id)
-	}
-	return out
-}
-
-func isJSONObject(raw string) bool {
-	var decoded map[string]any
-	if err := json.Unmarshal([]byte(strings.TrimSpace(raw)), &decoded); err != nil {
-		return false
-	}
-	return decoded != nil
 }
 
 func SandboxHasAgentTag(session *Sandbox, agentID string) bool {
