+++ pkg/model/agent_model_test.go
@@ -23,22 +23,6 @@ func TestNormalizeAgentKindPiAliases(t *testing.T) {
 	}
 }
 
-func TestNormalizeAgentDefinitionAcceptsPiAndRejectsUnknownProvider(t *testing.T) {
-	definition := AgentDefinition{ID: "pi-agent", Name: "reviewer", Provider: "pi-agent", Model: " openai/gpt-5.4 ", ProjectID: "project-1", AgentName: "reviewer"}
-	normalized, err := NormalizeAgentDefinition(definition, false)
-	if err != nil {
-		t.Fatalf("NormalizeAgentDefinition returned error: %v", err)
-	}
-	if normalized.Provider != "pi" || normalized.Model != "openai/gpt-5.4" {
-		t.Fatalf("normalized definition = %#v", normalized)
-	}
-
-	definition.Provider = "unknown"
-	if _, err := NormalizeAgentDefinition(definition, false); err == nil {
-		t.Fatal("NormalizeAgentDefinition accepted an unknown provider")
-	}
-}
-
 func TestProjectOwnershipJSONKeepsHistoricalFieldNames(t *testing.T) {
 	payload := struct {
 		Agent     AgentDefinition  `json:"agent"`
