+++ pkg/volumes/normalize_test.go
@@ -1,16 +1,18 @@
-package model
+package volumes
 
 import (
 	"strings"
 	"testing"
+
+	domain "agent-compose/pkg/model"
 )
 
 func TestNormalizeVolumeMountSpecsAllowsSameSourceMultipleTargetsAndNestedReadOnly(t *testing.T) {
-	items, err := NormalizeVolumeMountSpecs([]VolumeMountSpec{
+	items, err := NormalizeMountSpecs([]domain.VolumeMountSpec{
 		{Source: "cache", Target: "/mnt/cache-a"},
 		{Source: "cache", Target: "/mnt/cache-b"},
-		{Type: VolumeMountTypeVolume, Source: "nested-cache", Target: "/mnt/nested/parent/child", ReadOnly: true},
-		{Type: VolumeMountTypeBind, Source: "./logs", Target: "/mnt/logs"},
+		{Type: domain.VolumeMountTypeVolume, Source: "nested-cache", Target: "/mnt/nested/parent/child", ReadOnly: true},
+		{Type: domain.VolumeMountTypeBind, Source: "./logs", Target: "/mnt/logs"},
 	})
 	if err != nil {
 		t.Fatalf("NormalizeVolumeMountSpecs returned error: %v", err)
@@ -18,13 +20,13 @@ func TestNormalizeVolumeMountSpecsAllowsSameSourceMultipleTargetsAndNestedReadOn
 	if len(items) != 4 {
 		t.Fatalf("normalized item count = %d, want 4: %#v", len(items), items)
 	}
-	if items[0].Type != VolumeMountTypeVolume || items[1].Type != VolumeMountTypeVolume {
+	if items[0].Type != domain.VolumeMountTypeVolume || items[1].Type != domain.VolumeMountTypeVolume {
 		t.Fatalf("default volume mount types = %#v", items[:2])
 	}
 	if items[2].Target != "/mnt/nested/parent/child" || !items[2].ReadOnly {
 		t.Fatalf("nested read-only mount = %#v", items[2])
 	}
-	if items[3].Type != VolumeMountTypeBind || items[3].Source != "./logs" {
+	if items[3].Type != domain.VolumeMountTypeBind || items[3].Source != "./logs" {
 		t.Fatalf("bind mount = %#v", items[3])
 	}
 }
@@ -32,12 +34,12 @@ func TestNormalizeVolumeMountSpecsAllowsSameSourceMultipleTargetsAndNestedReadOn
 func TestNormalizeVolumeMountSpecsRejectsInvalidTargetsAndSources(t *testing.T) {
 	tests := []struct {
 		name string
-		spec []VolumeMountSpec
+		spec []domain.VolumeMountSpec
 		want string
 	}{
 		{
 			name: "duplicate cleaned target",
-			spec: []VolumeMountSpec{
+			spec: []domain.VolumeMountSpec{
 				{Source: "cache-a", Target: "/mnt/cache"},
 				{Source: "cache-b", Target: "/mnt/cache/."},
 			},
@@ -45,28 +47,28 @@ func TestNormalizeVolumeMountSpecsRejectsInvalidTargetsAndSources(t *testing.T) 
 		},
 		{
 			name: "relative target",
-			spec: []VolumeMountSpec{{Source: "cache", Target: "mnt/cache"}},
+			spec: []domain.VolumeMountSpec{{Source: "cache", Target: "mnt/cache"}},
 			want: `volume mount target "mnt/cache" must be absolute`,
 		},
 		{
 			name: "empty named volume source",
-			spec: []VolumeMountSpec{{Source: "", Target: "/mnt/cache"}},
+			spec: []domain.VolumeMountSpec{{Source: "", Target: "/mnt/cache"}},
 			want: "volume mount source: volume name is required",
 		},
 		{
 			name: "empty bind source",
-			spec: []VolumeMountSpec{{Type: VolumeMountTypeBind, Source: "", Target: "/mnt/logs"}},
+			spec: []domain.VolumeMountSpec{{Type: domain.VolumeMountTypeBind, Source: "", Target: "/mnt/logs"}},
 			want: "bind mount source is required",
 		},
 		{
 			name: "unsupported type",
-			spec: []VolumeMountSpec{{Type: "tmpfs", Source: "cache", Target: "/mnt/cache"}},
+			spec: []domain.VolumeMountSpec{{Type: "tmpfs", Source: "cache", Target: "/mnt/cache"}},
 			want: `volume mount type "tmpfs" is not supported`,
 		},
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			_, err := NormalizeVolumeMountSpecs(tt.spec)
+			_, err := NormalizeMountSpecs(tt.spec)
 			if err == nil {
 				t.Fatal("NormalizeVolumeMountSpecs returned nil error")
 			}
@@ -78,16 +80,16 @@ func TestNormalizeVolumeMountSpecsRejectsInvalidTargetsAndSources(t *testing.T) 
 }
 
 func TestNormalizeSessionVolumeMountsKeepsValidReadOnlyNestedMounts(t *testing.T) {
-	items := NormalizeSandboxVolumeMounts([]SandboxVolumeMount{
+	items := NormalizeSandboxMounts([]domain.SandboxVolumeMount{
 		{ID: " mount-a ", Type: " VOLUME ", Source: " cache ", Target: "/mnt/nested/../cache", ReadOnly: true, HostPath: " /host/cache "},
-		{Type: VolumeMountTypeVolume, Source: "missing-target", Target: "", HostPath: "/host/missing"},
-		{Type: VolumeMountTypeVolume, Source: "missing-host", Target: "/mnt/missing"},
+		{Type: domain.VolumeMountTypeVolume, Source: "missing-target", Target: "", HostPath: "/host/missing"},
+		{Type: domain.VolumeMountTypeVolume, Source: "missing-host", Target: "/mnt/missing"},
 	})
 	if len(items) != 1 {
 		t.Fatalf("normalized session mounts = %#v, want one valid mount", items)
 	}
 	item := items[0]
-	if item.ID != "mount-a" || item.Type != VolumeMountTypeVolume || item.Source != "cache" ||
+	if item.ID != "mount-a" || item.Type != domain.VolumeMountTypeVolume || item.Source != "cache" ||
 		item.Target != "/mnt/cache" || !item.ReadOnly || item.HostPath != "/host/cache" {
 		t.Fatalf("normalized session mount = %#v", item)
 	}
